# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webstore Customer API",
        "description": "Dokumentasi endpoint API customer untuk autentikasi dan fitur terkait dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/api/account/addresses": {
            "get": {
                "tags": [
                    "Customer Address"
                ],
                "summary": "Daftar alamat customer",
                "operationId": "3a6b8c5aea5b122956b908f8e7e668fb",
                "responses": {
                    "200": {
                        "description": "Daftar alamat berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Daftar alamat berhasil diambil.",
                                    "data": [
                                        {
                                            "id": 11,
                                            "label": "Rumah",
                                            "recipient_name": "Budi Santoso",
                                            "recipient_phone": "08123456789",
                                            "address_line1": "Jl. Merdeka No. 1",
                                            "province_label": "JAWA BARAT",
                                            "city_label": "KOTA BANDUNG",
                                            "district": "COBLONG",
                                            "is_default": true
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer Address"
                ],
                "summary": "Tambah alamat customer",
                "operationId": "65c3ea9410fb52104048fafbb048886e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_name",
                                    "recipient_phone",
                                    "address_line1",
                                    "province_label",
                                    "province_id",
                                    "city_label",
                                    "city_id",
                                    "district"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Rumah",
                                        "nullable": true
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "recipient_phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address_line1": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1"
                                    },
                                    "address_line2": {
                                        "type": "string",
                                        "example": "Dekat masjid",
                                        "nullable": true
                                    },
                                    "province_label": {
                                        "type": "string",
                                        "example": "JAWA BARAT"
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "example": 9
                                    },
                                    "city_label": {
                                        "type": "string",
                                        "example": "KOTA BANDUNG"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 501
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "COBLONG"
                                    },
                                    "district_lion": {
                                        "type": "string",
                                        "example": "COBLONG",
                                        "nullable": true
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "40132",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Indonesia",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Alamat pengiriman utama",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alamat berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Alamat berhasil ditambahkan.",
                                    "data": {
                                        "id": 11
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data tidak valid.",
                                    "errors": {
                                        "field": [
                                            "Field wajib diisi."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/account/addresses/{addressId}/default": {
            "post": {
                "tags": [
                    "Customer Address"
                ],
                "summary": "Set alamat default",
                "operationId": "b7572cf57700565709b949019446d631",
                "parameters": [
                    {
                        "name": "addressId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alamat default berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Alamat default berhasil diperbarui."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/account/addresses/options/provinces": {
            "get": {
                "tags": [
                    "Customer Address"
                ],
                "summary": "Opsi provinsi",
                "operationId": "5718b0b101ca31cf8e57ccd20bdd0e2a",
                "responses": {
                    "200": {
                        "description": "Daftar provinsi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 9
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "Jawa Barat"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 9,
                                        "label": "Jawa Barat"
                                    },
                                    {
                                        "id": 31,
                                        "label": "DKI Jakarta"
                                    }
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}